Added uniform support

+++ VEngine/src/events/MouseEvents.h
@@ -7,7 +7,7 @@ namespace vengine
 	class MouseMoveEvent : public Event
 	{
 	public:
-		MouseMoveEvent(const float mouse_x, const float mouse_y) : m_mouse_x(mouse_x), m_mouse_y(mouse_y)
+		MouseMoveEvent(const double mouse_x, const double mouse_y) : m_mouse_x(mouse_x), m_mouse_y(mouse_y)
 		{
 			m_event_type = EventType::MOUSE_MOVED;
 		}
